Updated Accessories Deletion

+++ src/app/api/admin/products/deleteAccessory/[accessoryId]/route.ts
@@ -2,7 +2,7 @@ import pool from "@/lib/db";
 
 export async function DELETE(req: Request, { params }: { params: { accessoryId: string } }) {
 
-  const { accessoryId } = params;
+  const { accessoryId } = await params;
 
   if (!accessoryId) {
     return new Response(JSON.stringify({ message: 'Accessory Id is required.' }), {
@@ -11,6 +11,19 @@ export async function DELETE(req: Request, { params }: { params: { accessoryId: 
   }
 
   try {
+    const checkQuery = 'SELECT COUNT(*) FROM products_accessories WHERE accessory_id = $1';
+    const checkValues = [accessoryId];
+
+    const checkResult = await pool.query(checkQuery, checkValues);
+
+    if (parseInt(checkResult.rows[0].count) > 0) {
+      return new Response(
+        JSON.stringify(
+          { message: 'This accessory is currently in use by one or more products. Please remove it from all products before proceeding with the deletion.' }),
+          { status: 400 }
+      );
+    }
+
     const query = 'DELETE FROM accessories WHERE accessory_id = $1 RETURNING *';
     const values = [accessoryId];
